api/src/main/post/requirements.txt
fastapi==0.95.1
SQLAlchemy==2.0.9
uvicorn==0.21.1
psycopg2-binary==2.9.6
fastapi-azure-auth==4.0.0
python-multipart==0.0.6
azure-storage-blob==12.16.0
